+++ internal/ui/cards.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"time"
 
-	zone "github.com/alexanderbh/bubblezone/v2"
-	"github.com/charmbracelet/lipgloss/v2"
+	"charm.land/lipgloss/v2"
+	zone "github.com/steipete/tmuxwatch/internal/zone"
 
 	"github.com/steipete/tmuxwatch/internal/tmux"
 )